Changed type

+++ src/main/groovy/Config.groovy
@@ -1,4 +1,4 @@
 class Config {
     public String url;
-    public long checkIntervalSec;
+    public int checkIntervalSec;
 }
